Allow passing in unbound modules into other modules
This change enables a `nn.Sequential` type pattern.

(though it's still confusing that you have to separately assign
submodules in `setup` -- #864 explores automating this somewhat.)

The original code was designed to allow using unbound modules as
"module templates" which you could bound arbitrarily in your tree,
but if people want that after this PR, they can always pass in lambdas.