Increase shuffle buffer size
A shuffle size smaller than the input data might have unexpected effects on the training process and thereby the final model quality, especially for heterogeneous and sharded datasets. Since most of the WMT datasets contain several millions of examples, it should be safer to set it to a higher number like this.