Replaces link to README with link to Github repo
The same README is shown as lower part of that Github repo. Having the link instead of `README` makes it more visible where the link is going. And it's a nice short link so there's no need to hide it behind a link text.

Finally, I often find it useful to navigate to Flax via `flax.readth...` (w/ Chrome autocomplete) - so it's nice having a link going from the to the Github repo.

(I know there is a "edit on Github link" - which I used for this edit -, but I don't usually want to edit anything, but rather browse the source or search things.)