Add nowrap decorator for module methods.
User-defined module methods are automatically wrapped with a state
management wrapper and, if profiling is enabled, with named_call.

This can interfere with constructor helper methods when they are
wrapped by named_call automatically.

Additionally for advanced cases of overriding e.g. the base param
method, one does not want to wrap such a method with the state
management wrapper.

So a @nn.nowrap method is introduced to prevent automatic wrapping
in those advanced usecases where it is detrimental.