Bringing changes in https://github.com/google/flax/pull/2955 internally.
"This PR aims to minimize the number of configuration files for the project. Instead of a separate .style.yapf, mypy.ini, pytest.ini, pytype.cfg and setup.cfg, we now would have all the configurations within the pyproject.toml file leading to a overall minimal code structure. I've also updated the .pre-commit-config.yaml file with the addition of a "check-toml" hook."

PiperOrigin-RevId: 521600833