[nnx] Some small documentation suggestions.
These are based on a quick read of the "basics" documentation. I wish I had a better way
to suggest changes. In some places are left "NOTE:" not because I propose to include
those changes but because I wanted to raise some questions. Perhaps we can discuss them
and agree whether to drop or to address them.