Add scale option to spectral normalization.
One can increase the expressivity while retaining some benefits
of spectral norm by parameterizing the spectral norm with
a single scale parameter. This forces the spectral norm to be
independent of the dimensionality of the weight matrix.
This was originally proposed in Appendix E of [1] and recently
applied to transformer models in [2].

[1] Takeru Miyato, Toshiki Kataoka, Masanori Koyama, and Yuichi Yoshida, 'Spectral Normalization for Generative Adversarial Networks', International Conference on Learning Representations (ICLR), 2018.
[2] Shuangfei Zhai, Tatiana Likhomanenko, Etai Littwin, Dan Busbridge, Jason Ramapuram, Yizhe Zhang, Jiatao Gu, and Joshua M. Susskind, 'Stabilizing Transformer Training by Preventing Attention Entropy Collapse'. International Conference on Machine Learning (ICML), 2023.