Fix bug when raising ScopeParamNotFoundError.
The order of 'actual value shape' vs. 'expected value shape based on initializer' was mixed up... the actual value should come first and the initialized-based expectation second. This CL changes order to be correct.

This removes a point of confusion when interpreting the error message when raised (expected vs. actual).

PiperOrigin-RevId: 799698584